requirements-dev.txt
# Test dependencies
pytest>=8.0.0
pytest-asyncio>=0.25.2
pytest-cov>=4.1.0
pytest-mock>=3.12.0
PyPDF2>=3.0.0

# Classifier dependencies
google-generativeai>=0.3.2
docling>=1.0.0
torch>=2.0.0  # Required for Docling

# Gmail API dependencies
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
simplegmail>=3.1.0

# Google Cloud dependencies
google-cloud-aiplatform>=1.0.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
click>=8.1.0  # Required for CLI tools
rich>=10.0.0  # Required for CLI UI
questionary>=2.0.0  # Required for interactive prompts
pydantic>=2.0.0

# Type stubs
types-PyYAML
types-python-dateutil

# Additional dependencies
types-Pillow==10.1.0.2
types-requests==2.31.0.20240106 